Development: Allow for debian packages in stock converter.

+++ development/stock-linux-json.py
@@ -3,10 +3,11 @@ import lzma
 import os
 import subprocess
 import tempfile
-from typing import List, Dict
+from typing import List, Dict, Optional
 
 import requests
 import rpmfile
+from debian import debfile
 
 DWARF2JSON = './dwarf2json'
 
@@ -22,32 +23,65 @@ class Downloader:
             files_for_processing = self.download_list(url_list)
             self.process_files(files_for_processing)
             for fname in files_for_processing.values():
-                os.unlink(fname)
+                if fname:
+                    os.unlink(fname)
 
     def download_list(self, urls: List[str]) -> Dict[str, str]:
         processed_files = {}
         for url in urls:
             print(" - Downloading {}".format(url))
             data = requests.get(url)
-            with tempfile.NamedTemporaryFile() as rpmdata:
-                rpmdata.write(data.content)
-                rpmdata.seek(0)
-                rpm = rpmfile.RPMFile(fileobj = rpmdata)
-                for member in rpm.getmembers():
-                    if 'vmlinux' in member.name or 'System.map' in member.name:
-                        print(" - Extracting {}".format(member.name))
-                        extracted = rpm.extractfile(member)
-                        break
-                with tempfile.NamedTemporaryFile(
-                        delete = False, prefix = 'vmlinux' if 'vmlinux' in member.name else 'System.map') as output:
-                    print(" - Writing to {}".format(output.name))
-                    output.write(extracted.read())
-                    processed_files[url] = output.name
+            with tempfile.NamedTemporaryFile() as archivedata:
+                archivedata.write(data.content)
+                archivedata.seek(0)
+                if url.endswith('.rpm'):
+                    processed_files[url] = self.process_rpm(archivedata)
+                elif url.endswith('.deb'):
+                    processed_files[url] = self.process_deb(archivedata)
+
         return processed_files
+
+    def process_rpm(self, archivedata) -> Optional[str]:
+        rpm = rpmfile.RPMFile(fileobj = archivedata)
+        member = None
+        extracted = None
+        for member in rpm.getmembers():
+            if 'vmlinux' in member.name or 'System.map' in member.name:
+                print(" - Extracting {}".format(member.name))
+                extracted = rpm.extractfile(member)
+                break
+        if not member or not extracted:
+            return None
+        with tempfile.NamedTemporaryFile(delete = False,
+                                         prefix = 'vmlinux' if 'vmlinux' in member.name else 'System.map') as output:
+            print(" - Writing to {}".format(output.name))
+            output.write(extracted.read())
+        return output.name
+
+    def process_deb(self, archivedata) -> Optional[str]:
+        deb = debfile.DebFile(fileobj = archivedata)
+        member = None
+        extracted = None
+        for member in deb.data.tgz().getmembers():
+            if member.name.endswith('vmlinux') or 'System.map' in member.name:
+                print(" - Extracting {}".format(member.name))
+                extracted = deb.data.get_file(member.name)
+                break
+        if not member or not extracted:
+            return None
+        with tempfile.NamedTemporaryFile(delete = False,
+                                         prefix = 'vmlinux' if 'vmlinux' in member.name else 'System.map') as output:
+            print(" - Writing to {}".format(output.name))
+            output.write(extracted.read())
+        return output.name
 
     def process_files(self, named_files: Dict[str, str]):
         """Runs the dwarf2json binary across the files"""
         print("Processing Files...")
+        for i in named_files:
+            if named_files[i] is None:
+                print("FAILURE: None encountered for {}".format(i))
+                return
         args = [DWARF2JSON, 'linux']
         output_filename = 'unknown-kernel.json'
         for named_file in named_files:
